Fix merge errors
