+++ api/src/test/java/com/example/demo/service/CartListServiceTest.java
@@ -1,7 +1,7 @@
 package com.example.demo.service;
 
-import com.example.demo.Repository.CartListRepository;
-import com.example.demo.Service.CartListService;
+import com.example.demo.repository.CartListRepository;
+import com.example.demo.service.CartListService;
 
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
