+++ api/src/test/java/com/example/demo/service/OrderListServiceTest.java
@@ -1,7 +1,7 @@
 package com.example.demo.service;
 
-import com.example.demo.Repository.OrderListRepository;
-import com.example.demo.Service.OrderListService;
+import com.example.demo.repository.OrderListRepository;
+import com.example.demo.service.OrderListService;
 
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
