+++ api/src/test/java/com/example/demo/service/ProductServiceTest.java
@@ -1,7 +1,7 @@
 package com.example.demo.service;
 
-import com.example.demo.Repository.ProductRepository;
-import com.example.demo.Service.ProductService;
+import com.example.demo.repository.ProductRepository;
+import com.example.demo.service.ProductService;
 
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
