+++ api/src/test/java/com/example/demo/service/StoreServiceTest.java
@@ -1,7 +1,7 @@
 package com.example.demo.service;
 
-import com.example.demo.Repository.StoreRepository;
-import com.example.demo.Service.StoreService;
+import com.example.demo.repository.StoreRepository;
+import com.example.demo.service.StoreService;
 
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
