+++ api/src/test/java/com/example/demo/service/UserServiceTest.java
@@ -1,7 +1,7 @@
 package com.example.demo.service;
 
-import com.example.demo.Repository.UserRepository;
-import com.example.demo.Service.UserService;
+import com.example.demo.repository.UserRepository;
+import com.example.demo.service.UserService;
 
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
